refactor: 추가 리팩터링
